requirements.txt
# Core data processing and analysis
pandas>=1.5.0
numpy>=1.21.0

# Geospatial distance calculations
geopy>=2.3.0

# Machine learning and preprocessing
scikit-learn>=1.3.0
joblib>=1.2.0
catboost>=1.2.0

# Deep learning
torch>=1.13.0
tensorflow>=2.10.0
keras-tcn>=3.9.0

# Data visualization
matplotlib

# Web automation / scraping (lineups, subs)
selenium>=4.10.0

# Progress bar visualization
tqdm>=4.64.0

# Optional: High-performance data format support
# Uncomment if you want to use Parquet format for faster I/O
# pyarrow>=10.0.0
# fastparquet>=0.8.0

# Development and testing (optional)
# pytest>=7.0.0
# black>=22.0.0
# flake8>=5.0.0
